+++ tests/framework/rest/SerializerTest.php
@@ -3,6 +3,7 @@
 namespace yiiunit\framework\rest;
 
 use yii\base\Model;
+use yii\data\ArrayDataProvider;
 use yii\rest\Serializer;
 use yiiunit\TestCase;
 
@@ -29,7 +30,23 @@ class SerializerTest extends TestCase
 
     public function testSerializeModelErrors()
     {
-        // TODO
+        $serializer = new Serializer();
+        $model = new TestModel();
+
+        $model->addError('field1', 'Test error');
+        $model->addError('field2', 'Multiple error 1');
+        $model->addError('field2', 'Multiple error 2');
+
+        $this->assertEquals([
+            [
+                'field' => 'field1',
+                'message' => 'Test error',
+            ],
+            [
+                'field' => 'field2',
+                'message' => 'Multiple error 1',
+            ]
+        ], $serializer->serialize($model));
     }
 
     public function testSerializeModelData()
@@ -118,12 +135,88 @@ class SerializerTest extends TestCase
         ], $serializer->serialize($model));
     }
 
-    public function testSerializeDataProvider()
+    public function dataProviderSerializeDataProvider()
     {
-        // TODO
+        return [
+            [
+                new ArrayDataProvider([
+                    'allModels' => [
+                        ['id' => 1, 'username' => 'Bob'],
+                        ['id' => 2, 'username' => 'Tom']
+                    ],
+                    'pagination' => [
+                        'route' => '/',
+                    ],
+                ]),
+                [
+                    ['id' => 1, 'username' => 'Bob'],
+                    ['id' => 2, 'username' => 'Tom']
+                ]
+            ],
+            [
+                new ArrayDataProvider([
+                    'allModels' => [
+                        ['id' => 1, 'username' => 'Bob'],
+                        ['id' => 2, 'username' => 'Tom']
+                    ],
+                    'pagination' => [
+                        'route' => '/',
+                        'pageSize' => 1,
+                        'page' => 0
+                    ],
+                ]),
+                [
+                    ['id' => 1, 'username' => 'Bob'],
+                ]
+            ],
+            [
+                new ArrayDataProvider([
+                    'allModels' => [
+                        ['id' => 1, 'username' => 'Bob'],
+                        ['id' => 2, 'username' => 'Tom']
+                    ],
+                    'pagination' => [
+                        'route' => '/',
+                        'pageSize' => 1,
+                        'page' => 1
+                    ],
+                ]),
+                [
+                    ['id' => 2, 'username' => 'Tom']
+                ]
+            ],
+            /*[
+                new ArrayDataProvider([
+                    'allModels' => [
+                        new \DateTime('2000-01-01'),
+                    ],
+                    'pagination' => [
+                        'route' => '/',
+                    ],
+                ]),
+                [
+                    [
+                        'date' => '2000-01-01 00:00:00.000000',
+                        'timezone_type' => 3,
+                        'timezone' => 'UTC',
+                    ],
+                ]
+            ],*/
+        ];
     }
 
+    /**
+     * @dataProvider dataProviderSerializeDataProvider
+     *
+     * @param \yii\data\DataProviderInterface $dataProvider
+     * @param array $expectedResult
+     */
+    public function testSerializeDataProvider($dataProvider, $expectedResult)
+    {
+        $serializer = new Serializer();
 
+        $this->assertEquals($expectedResult, $serializer->serialize($dataProvider));
+    }
 }
 
 class TestModel extends Model